adding some intelligent (semi-intelligent) renaming such that variables that contain Lustre keywords are renamed and the layout does not lose them.
For example:

intruder becomes intruder_ (the ID starts with a Lustre keyword 'int')
realistic becomes realistic_ (the ID starts with a Lustre keyword
'real')
node becomes node_ (the ID starts with, and equals, a Lustre keyword
'node)

These new, renamed variables are relayed to the user. Additionally, the
renamed variables are output to the console. 

I am unsure as of right now of a better way to make the user aware of
the renaming.