Delete Generated Artifacts
For issue #55

Keep the src-gen and xtend-gen directories, since they are required on the build path.

Modify the mwe2 to not delete the .keep files from the src-gen directories, so that those directories will continue to hang around.

Modify the .gitignore so that we are not told about generated language files.